+++ cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @kilocode/cli
 
+## 0.26.0
+
+### Minor Changes
+
+- [#5330](https://github.com/Kilo-Org/kilocode/pull/5330) [`957df89`](https://github.com/Kilo-Org/kilocode/commit/957df89a92d951c409952e16948694488abce474) Thanks [@qbiecom](https://github.com/qbiecom)! - Added OpenAI Compatible (Responses) provider
+
+### Patch Changes
+
+- [#5302](https://github.com/Kilo-Org/kilocode/pull/5302) [`292c885`](https://github.com/Kilo-Org/kilocode/commit/292c885fc73424b2fbe2c36311f57662b495770e) Thanks [@marius-kilocode](https://github.com/marius-kilocode)! - Keep "Thinking..." indicator visible in CLI after checkpoint saves and API calls complete
+
+## 0.25.1
+
+### Patch Changes
+
+- [#5250](https://github.com/Kilo-Org/kilocode/pull/5250) [`96c2b39`](https://github.com/Kilo-Org/kilocode/commit/96c2b391744beb5b1a1665638587cc67706f6773) Thanks [@marius-kilocode](https://github.com/marius-kilocode)! - Fix intermittent issue where list_files and search_files tools show "Total: 0 items" or "Found: 0 matches"
+
+- [#5227](https://github.com/Kilo-Org/kilocode/pull/5227) [`12d96fa`](https://github.com/Kilo-Org/kilocode/commit/12d96fa41da4ce1ae3dce03d7180f59f1e154ff8) Thanks [@marius-kilocode](https://github.com/marius-kilocode)! - Replace verbose "(git worktree)" indicator with compact "⎇" symbol in CLI status bar and terminal title
+
 ## 0.25.0
 
 ### Minor Changes